Fix displaying format for float

+++ types.go
@@ -10,7 +10,7 @@ const (
 type transactionPayload struct {
 	UserID          string  `json:"user_id"`
 	Currency        string  `json:"currency"`
-	Amount          float32 `json:"amount"`
+	Amount          float64 `json:"amount"`
 	TimePlaced      string  `json:"time_placed"`
 	TransactionType string  `json:"type"`
 }
@@ -19,7 +19,7 @@ type transactionModel struct {
 	tableName         struct{} `pg:"transactions"`
 	ID                int      `json:"id" pg:"id"`
 	WalletID          int      `json:"wallet_id"`
-	Amount            float32  `json:"amount"`
+	Amount            float64  `json:"amount"`
 	TimePlaced        string   `json:"time_placed"`
 	TransactionTypeID int      `json:"transaction_type_id"`
 	Status            int      `json:"status"`
@@ -29,8 +29,8 @@ type walletModel struct {
 	tableName  struct{} `pg:"wallets"`
 	ID         int      `json:"id" pg:"id"`
 	UserID     string   `json:"user_id"`
-	CurrencyID float32  `json:"currency_id"`
-	Balance    float32  `json:"balance"`
+	CurrencyID float64  `json:"currency_id"`
+	Balance    float64  `json:"balance"`
 }
 
 type currencyModel struct {
